requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
uvicorn-worker>=0.2.0
gunicorn>=22.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic-settings>=2.0.0
httpx>=0.27.0
anthropic>=0.40.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pytest>=8.0.0
